+++ beacon-chain/forkchoice/protoarray/unrealized_justification_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	forkchoicetypes "github.com/prysmaticlabs/prysm/beacon-chain/forkchoice/types"
+	"github.com/prysmaticlabs/prysm/config/features"
 	"github.com/prysmaticlabs/prysm/config/params"
 	types "github.com/prysmaticlabs/prysm/consensus-types/primitives"
 	"github.com/prysmaticlabs/prysm/testing/require"
@@ -247,4 +248,91 @@ func TestStore_ForkNextEpoch(t *testing.T) {
 	// nodes[8] = D since it's late!
 	require.Equal(t, uint64(0), f.store.nodes[8].weight)
 	require.Equal(t, uint64(100), f.store.nodes[7].weight)
+}
+
+func TestStore_PullTips_Heuristics(t *testing.T) {
+	resetCfg := features.InitWithReset(&features.Flags{
+		PullTips: true,
+	})
+	defer resetCfg()
+	ctx := context.Background()
+	t.Run("Current epoch is justified", func(tt *testing.T) {
+		f := setup(1, 1)
+		st, root, err := prepareForkchoiceState(ctx, 65, [32]byte{'p'}, [32]byte{}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		f.store.nodes[1].unrealizedJustifiedEpoch = types.Epoch(2)
+		driftGenesisTime(f, 66, 0)
+
+		st, root, err = prepareForkchoiceState(ctx, 66, [32]byte{'h'}, [32]byte{'p'}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		require.Equal(tt, types.Epoch(2), f.store.nodes[2].unrealizedJustifiedEpoch)
+		require.Equal(tt, types.Epoch(1), f.store.nodes[2].unrealizedFinalizedEpoch)
+	})
+
+	t.Run("Previous Epoch is justified and too early for current", func(tt *testing.T) {
+		f := setup(1, 1)
+		st, root, err := prepareForkchoiceState(ctx, 95, [32]byte{'p'}, [32]byte{}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		f.store.nodes[1].unrealizedJustifiedEpoch = types.Epoch(2)
+		driftGenesisTime(f, 96, 0)
+
+		st, root, err = prepareForkchoiceState(ctx, 96, [32]byte{'h'}, [32]byte{'p'}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		require.Equal(tt, types.Epoch(2), f.store.nodes[2].unrealizedJustifiedEpoch)
+		require.Equal(tt, types.Epoch(1), f.store.nodes[2].unrealizedFinalizedEpoch)
+	})
+	t.Run("Previous Epoch is justified and not too early for current", func(tt *testing.T) {
+		f := setup(1, 1)
+		st, root, err := prepareForkchoiceState(ctx, 95, [32]byte{'p'}, [32]byte{}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		f.store.nodes[1].unrealizedJustifiedEpoch = types.Epoch(2)
+		driftGenesisTime(f, 127, 0)
+
+		st, root, err = prepareForkchoiceState(ctx, 127, [32]byte{'h'}, [32]byte{'p'}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		// Check that the justification point is not the parent's.
+		// This tests that the heuristics in pullTips did not apply and
+		// the test continues to compute a bogus unrealized
+		// justification
+		require.Equal(tt, types.Epoch(1), f.store.nodes[2].unrealizedJustifiedEpoch)
+	})
+	t.Run("Block from previous Epoch", func(tt *testing.T) {
+		f := setup(1, 1)
+		st, root, err := prepareForkchoiceState(ctx, 94, [32]byte{'p'}, [32]byte{}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		f.store.nodes[1].unrealizedJustifiedEpoch = types.Epoch(2)
+		driftGenesisTime(f, 96, 0)
+
+		st, root, err = prepareForkchoiceState(ctx, 95, [32]byte{'h'}, [32]byte{'p'}, [32]byte{}, 1, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		// Check that the justification point is not the parent's.
+		// This tests that the heuristics in pullTips did not apply and
+		// the test continues to compute a bogus unrealized
+		// justification
+		require.Equal(tt, types.Epoch(1), f.store.nodes[2].unrealizedJustifiedEpoch)
+	})
+	t.Run("Previous Epoch is not justified", func(tt *testing.T) {
+		f := setup(1, 1)
+		st, root, err := prepareForkchoiceState(ctx, 128, [32]byte{'p'}, [32]byte{}, [32]byte{}, 2, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		driftGenesisTime(f, 129, 0)
+
+		st, root, err = prepareForkchoiceState(ctx, 129, [32]byte{'h'}, [32]byte{'p'}, [32]byte{}, 2, 1)
+		require.NoError(tt, err)
+		require.NoError(tt, f.InsertNode(ctx, st, root))
+		// Check that the justification point is not the parent's.
+		// This tests that the heuristics in pullTips did not apply and
+		// the test continues to compute a bogus unrealized
+		// justification
+		require.Equal(tt, types.Epoch(2), f.store.nodes[2].unrealizedJustifiedEpoch)
+	})
 }
